Allow configuration of the color palette.
The palette is passed to the row generator and not when creating the
Data struct to avoid me having to think about lifetimes. Much of the
messaging code could be refactored, so the palette improvement can be
done there.

Also adds four palettes, that work fine with a dark background but not
fine with a light background. This is in line with the rest of the app
at the moment.