Implement basic readline keybindings.
This lets us press CTRL-W to cut a word and CTRL-U to cut to the start
of the line just like default readline bindings (emacs).

None of the other bindings are feasible at the moment because we are not
tracking cursor position. Unfortunately the handling for CTRL-W will
have to be changed once we track cursor position, but the logic should
be the same as the one implemented here so any changes should be trivial.