[skip ci] Fixed crash when emotes are disabled
I absolutely despise this solution, but it's a hotfix for now. There
should be a singleton object created before `MessageData` and
`RawMessageData`, while having access to both such that there is no code
duplication. That is off the top of my head, I'll come up with something
that will actually be good later.